tests: factor out zip and store helper functions
