+++ cmd/replacer/replace/replace_test.go
@@ -1,11 +1,7 @@
 package replace_test
 
 import (
-	"archive/tar"
-	"bytes"
-	"context"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -17,9 +13,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/replacer/protocol"
 	"github.com/sourcegraph/sourcegraph/cmd/replacer/replace"
-	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	"github.com/sourcegraph/sourcegraph/internal/store"
+	"github.com/sourcegraph/sourcegraph/internal/testutil"
 )
 
 func TestReplace(t *testing.T) {
@@ -56,7 +50,7 @@ func main() {
 `},
 	}
 
-	store, cleanup, err := newStore(files)
+	store, cleanup, err := testutil.NewStore(files)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -105,7 +99,7 @@ func TestReplace_badrequest(t *testing.T) {
 		},
 	}
 
-	store, cleanup, err := newStore(nil)
+	store, cleanup, err := testutil.NewStore(nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -149,45 +143,6 @@ func doReplace(u string, p *protocol.Request) (string, error) {
 	return string(body), err
 }
 
-func newStore(files map[string]string) (*store.Store, func(), error) {
-	buf := new(bytes.Buffer)
-	w := tar.NewWriter(buf)
-	for name, body := range files {
-		hdr := &tar.Header{
-			Name: name,
-			Mode: 0600,
-			Size: int64(len(body)),
-		}
-		if err := w.WriteHeader(hdr); err != nil {
-			return nil, nil, err
-		}
-		if _, err := w.Write([]byte(body)); err != nil {
-			return nil, nil, err
-		}
-	}
-	// git-archive usually includes a pax header we should ignore.
-	// use a body which matches a test case. Ensures we don't return this
-	// false entry as a result.
-	if err := addpaxheader(w, "Hello world\n"); err != nil {
-		return nil, nil, err
-	}
-
-	err := w.Close()
-	if err != nil {
-		return nil, nil, err
-	}
-	d, err := ioutil.TempDir("", "search_test")
-	if err != nil {
-		return nil, nil, err
-	}
-	return &store.Store{
-		FetchTar: func(ctx context.Context, repo gitserver.Repo, commit api.CommitID) (io.ReadCloser, error) {
-			return ioutil.NopCloser(bytes.NewReader(buf.Bytes())), nil
-		},
-		Path: d,
-	}, func() { os.RemoveAll(d) }, nil
-}
-
 func diff(b1, b2 string) (string, error) {
 	f1, err := ioutil.TempFile("", "search_test")
 	if err != nil {
@@ -217,13 +172,4 @@ func diff(b1, b2 string) (string, error) {
 		err = nil
 	}
 	return string(data), err
-}
-
-func addpaxheader(w *tar.Writer, body string) error {
-	hdr := &tar.Header{
-		Name:       "pax_global_header",
-		Typeflag:   tar.TypeXGlobalHeader,
-		PAXRecords: map[string]string{"somekey": body},
-	}
-	return w.WriteHeader(hdr)
 }
