+++ cmd/searcher/search/search_structural_test.go
@@ -2,14 +2,12 @@ package search
 
 import (
 	"context"
-	"io/ioutil"
-	"os"
 	"reflect"
 	"sort"
 	"testing"
 
 	"github.com/sourcegraph/sourcegraph/cmd/searcher/protocol"
-	"github.com/sourcegraph/sourcegraph/internal/store"
+	"github.com/sourcegraph/sourcegraph/internal/testutil"
 )
 
 // Tests that structural search correctly infers the Go matcher from the .go
@@ -27,11 +25,11 @@ func foo(real string) {}
 
 	includePatterns := []string{"main.go"}
 
-	zipData, err := createZip(input)
+	zipData, err := testutil.CreateZip(input)
 	if err != nil {
 		t.Fatal(err)
 	}
-	zf, cleanup, err := MockZipFileOnDisk(zipData)
+	zf, cleanup, err := testutil.TempZipFileOnDisk(zipData)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -79,11 +77,11 @@ func TestIncludePatterns(t *testing.T) {
 
 	includePatterns := []string{"a/b/c/foo.go", "bar.go"}
 
-	zipData, err := createZip(input)
+	zipData, err := testutil.CreateZip(input)
 	if err != nil {
 		t.Fatal(err)
 	}
-	zf, cleanup, err := MockZipFileOnDisk(zipData)
+	zf, cleanup, err := testutil.TempZipFileOnDisk(zipData)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -106,24 +104,4 @@ func TestIncludePatterns(t *testing.T) {
 	if !reflect.DeepEqual(got, want) {
 		t.Fatalf("got file matches %v, want %v", got, want)
 	}
-}
-
-func MockZipFileOnDisk(data []byte) (string, func(), error) {
-	z, err := store.MockZipFile(data)
-	if err != nil {
-		return "", nil, err
-	}
-	d, err := ioutil.TempDir("", "search_test")
-	if err != nil {
-		return "", nil, err
-	}
-	f, err := ioutil.TempFile(d, "search_zip")
-	if err != nil {
-		return "", nil, err
-	}
-	_, err = f.Write(z.Data)
-	if err != nil {
-		return "", nil, err
-	}
-	return f.Name(), func() { os.RemoveAll(d) }, nil
 }
